Refactor out generationVersion function to calculate version number
